pyproject.toml
[tool.poetry]
name = "cookieclicker"
version = "0.1.0"
description = ""
authors = ["gdoerksen <gramdoer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
PyAutoGUI = "^0.9.53"
numpy = "^1.23.3"
pandas = "^1.4.4"
Pillow = "^9.2.0"
opencv-contrib-python = "^4.6.0.66"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
